9/24 Foreign keys & voice chat handling
9/24
very confusing on how should I handle the voice-chat-leveling-system. Sole idea is whenever new one join, run a for loop with all other member in the vc and create a conn_line obj with them. conn_line  will be recorded inside a dictionary with a key of the member1.

added two other tables using foreign keys to clean up the original table